+++ AdjustSdkIm/AdjustSdkIm.h
@@ -29,3 +29,5 @@ FOUNDATION_EXPORT const unsigned char AdjustSdkImVersionString[];
 #import <AdjustSdkIm/ADJSessionFailure.h>
 #import <AdjustSdkIm/ADJAdRevenue.h>
 #import <AdjustSdkIm/ADJLinkResolution.h>
+#import <AdjustSdkIm/ADJPurchase.h>
+#import <AdjustSdkIm/ADJPurchaseVerificationResult.h>