+++ AdjustSdkStatic/AdjustSdk.h
@@ -21,3 +21,5 @@
 #import <AdjustSdk/ADJSessionFailure.h>
 #import <AdjustSdk/ADJAdRevenue.h>
 #import <AdjustSdk/ADJLinkResolution.h>
+#import <AdjustSdk/ADJPurchase.h>
+#import <AdjustSdk/ADJPurchaseVerificationResult.h>