+++ AdjustSdkWebBridge/AdjustSdkWebBridge.h
@@ -30,6 +30,8 @@ FOUNDATION_EXPORT const unsigned char AdjustSdkWebBridgeVersionString[];
 #import <AdjustSdkWebBridge/ADJSessionFailure.h>
 #import <AdjustSdkWebBridge/ADJAdRevenue.h>
 #import <AdjustSdkWebBridge/ADJLinkResolution.h>
+#import <AdjustSdkWebBridge/ADJPurchase.h>
+#import <AdjustSdkWebBridge/ADJPurchaseVerificationResult.h>
 
 // Exposing entire WebViewJavascriptBridge framework
 #import <AdjustSdkWebBridge/WebViewJavascriptBridge_JS.h>