+++ examples/AdjustExample-ObjC/AdjustExample-ObjC/AppDelegate.m
@@ -23,7 +23,7 @@
     
     // Change the log level.
     [adjustConfig setLogLevel:ADJLogLevelVerbose];
-    
+
     // Enable event buffering.
     // [adjustConfig setEventBufferingEnabled:YES];